% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-definition.R
\docType{class}
\name{tensorBag-class}
\alias{tensorBag-class}
\alias{tensorBag}
\title{定义一个张量包，这是张量在R中的基本单元}
\value{
返回值
}
\description{
定义一个张量包，这是张量在R中的基本单元
}
\section{Slots}{

\describe{
\item{\code{edgeCount}}{integer. 张量的阶数}

\item{\code{edges}}{integer.  张量的每一阶的维数，可以不一样}

\item{\code{edgeNames}}{character. 张量的初始名称，系统自动命名}

\item{\code{data}}{array. 张量的数据存储，在tensor.toolkit中使用array的type.}

\item{\code{dimNames}}{维度名称}

\item{\code{custNames}}{用户定义的维度名称}

\item{\code{nickNames}}{自动定义的维度名称}

\item{\code{useNick}}{是否使用自动定义的维度名称}

\item{\code{full_capacity}}{张量理论容量}

\item{\code{real_capacity}}{张量实际容量}

\item{\code{diff_capacity}}{张量差异容量}

\item{\code{useFill}}{是否需要填充}

\item{\code{fillData}}{填充数据}

\item{\code{edgeRank}}{阶排序}

\item{\code{dimRank}}{维排序}

\item{\code{fillValue}}{插补值}

\item{\code{valueName}}{值列名称 值名称}

\item{\code{dimExpTable}}{维度扩展表}

\item{\code{srcClass}}{源数据类型}

\item{\code{synChkTable}}{同步检查表}
}}

\examples{
tensorBag();
}
