% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-definition.R
\docType{class}
\name{tensorList-class}
\alias{tensorList-class}
\alias{tensorList}
\title{定义张量列表}
\value{
返回值
}
\description{
定义张量列表
}
\section{Slots}{

\describe{
\item{\code{data}}{list.张量列表数据，每个子类类型应该为tensorBag或tensorList}

\item{\code{itemType}}{character. 元素类型，一般为tensorBag或tensorList}

\item{\code{itemCount}}{integer. 元素个数}

\item{\code{listLevel}}{列表级次，默认为1}

\item{\code{listNames}}{列表名称}

\item{\code{is_detail}}{是否明细当listLevel为1,itemType为tensorBag为TRUE}
}}

\examples{
tensorList();
}
